dbms: fixed error with merges [#METR-12767].

+++ dbms/include/DB/Storages/MergeTree/BackgroundProcessingPool.h
@@ -101,7 +101,7 @@ public:
 	}
 
 
-	int getNumberOfThreads() const
+	size_t getNumberOfThreads() const
 	{
 		return size;
 	}
